revamp open in editor api and bring it into useStorybookApi
