+++ code/core/src/component-testing/components/InteractionsPanel.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 
 import { transparentize } from 'polished';
+import type { API } from 'storybook/manager-api';
 import { styled } from 'storybook/theming';
 
 import { type Call, type CallStates, type ControlStates } from '../../instrumenter/types';
@@ -46,6 +47,7 @@ interface InteractionsPanelProps {
   browserTestStatus?: CallStates;
   importPath?: string;
   canOpenInEditor?: boolean;
+  api: API;
 }
 
 const Container = styled.div(({ theme }) => ({
@@ -108,6 +110,7 @@ export const InteractionsPanel: React.FC<InteractionsPanelProps> = React.memo(
     browserTestStatus,
     importPath,
     canOpenInEditor,
+    api,
   }) {
     const filter = useAnsiToHtmlFilter();
     const hasRealInteractions = interactions.some((i) => i.id !== INTERNAL_RENDER_CALL_ID);
@@ -126,6 +129,7 @@ export const InteractionsPanel: React.FC<InteractionsPanelProps> = React.memo(
           onScrollToEnd={onScrollToEnd}
           importPath={importPath}
           canOpenInEditor={canOpenInEditor}
+          api={api}
         />
         <div aria-label="Interactions list">
           {interactions.map((call) => (
