+++ code/core/src/core-server/server-channel/open-in-editor-channel.ts
@@ -1,12 +1,7 @@
-import { join } from 'node:path';
-
 import type { Channel } from 'storybook/internal/channels';
-import { getProjectRoot } from 'storybook/internal/common';
 import type {
   OpenInEditorRequestPayload,
   OpenInEditorResponsePayload,
-  RequestData,
-  ResponseData,
 } from 'storybook/internal/core-events';
 import { OPEN_IN_EDITOR_REQUEST, OPEN_IN_EDITOR_RESPONSE } from 'storybook/internal/core-events';
 import { telemetry } from 'storybook/internal/telemetry';
