+++ code/core/src/manager-api/index.ts
@@ -1,7 +1,5 @@
 export * from './root';
 
-export { openInEditor } from './lib/open-in-editor';
-
 export { UniversalStore as experimental_UniversalStore } from '../shared/universal-store';
 export { useUniversalStore as experimental_useUniversalStore } from '../shared/universal-store/use-universal-store-manager';
 export { MockUniversalStore as experimental_MockUniversalStore } from '../shared/universal-store/mock';
