+++ code/core/src/manager-api/modules/notifications.ts
@@ -1,15 +1,6 @@
-import React from 'react';
-
-import {
-  OPEN_IN_EDITOR_RESPONSE,
-  type OpenInEditorResponsePayload,
-} from 'storybook/internal/core-events';
 import type { API_Notification } from 'storybook/internal/types';
 
-import { FailedIcon } from '@storybook/icons';
-
 import { partition } from 'es-toolkit/array';
-import { color } from 'storybook/theming';
 
 import type { ModuleFn } from '../lib/types';
 
@@ -35,7 +26,7 @@ export interface SubAPI {
   clearNotification: (id: string) => void;
 }
 
-export const init: ModuleFn = ({ store, provider }) => {
+export const init: ModuleFn = ({ store }) => {
   const api: SubAPI = {
     addNotification: (newNotification) => {
       store.setState(({ notifications }) => {
@@ -67,22 +58,5 @@ export const init: ModuleFn = ({ store, provider }) => {
   return {
     api,
     state,
-    init: async () => {
-      provider.channel?.on(OPEN_IN_EDITOR_RESPONSE, (payload: OpenInEditorResponsePayload) => {
-        if (payload.error !== null) {
-          api.addNotification({
-            id: 'open-in-editor-error',
-            content: {
-              headline: 'Failed to open in editor',
-              subHeadline:
-                payload.error ||
-                'Check the Storybook process on the command line for more details.',
-            },
-            icon: <FailedIcon color={color.negative} />,
-            duration: 8_000,
-          });
-        }
-      });
-    },
   };
 };
