+++ code/core/src/manager/globals/exports.ts
@@ -339,7 +339,6 @@ export default {
     'keyToSymbol',
     'merge',
     'mockChannel',
-    'openInEditor',
     'optionOrAltSymbol',
     'shortcutMatchesShortcut',
     'shortcutToHumanString',
